tierkreis_workers/aer_worker/pyproject.toml
[project]
name = "aer_worker"
version = "0.1.0"
description = "A collection of python tierkreis workers."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "tierkreis",
    "pytket-qiskit",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = [
  "*.py",
]
[tool.hatch.build.targets.wheel]
include = [
  "*.py",
]

[tool.uv.sources]
tierkreis = { workspace = true }

[project.scripts]
tkr_aer_worker = "main:main"
